move tests and fixtures of kube_client to a separate file
