+++ tests/integration/test_kube_orchestrator.py
@@ -3,19 +3,8 @@ import io
 import time
 import uuid
 from pathlib import PurePath
-from typing import (
-    Any,
-    AsyncIterator,
-    Awaitable,
-    Callable,
-    Dict,
-    Iterator,
-    Optional,
-    Sequence,
-)
-from unittest import mock
+from typing import Any, Awaitable, Callable, Dict, Iterator, Optional, Sequence
 
-import aiohttp
 import pytest
 from async_timeout import timeout
 from elasticsearch import AuthenticationException
@@ -45,17 +34,10 @@ from platform_api.orchestrator.job_request import (
     ContainerVolume,
 )
 from platform_api.orchestrator.kube_client import (
-    AlreadyExistsException,
-    DockerRegistrySecret,
-    Ingress,
-    IngressRule,
     KubeClient,
-    KubeClientException,
-    PodContainerStats,
     PodDescriptor,
     PodStatus,
     SecretRef,
-    Service,
     StatusException,
 )
 from platform_api.orchestrator.kube_orchestrator import JobStatusItemFactory
@@ -96,24 +78,6 @@ async def job_nginx(kube_orchestrator: KubeOrchestrator) -> MyJob:
     job_request = JobRequest.create(container)
     job = MyJob(orchestrator=kube_orchestrator, job_request=job_request)
     return job
-
-
-@pytest.fixture
-async def delete_job_later(
-    kube_orchestrator: KubeOrchestrator
-) -> AsyncIterator[Callable[[Job], Awaitable[None]]]:
-    jobs = []
-
-    async def _add_job(job: Job) -> None:
-        jobs.append(job)
-
-    yield _add_job
-
-    for job in jobs:
-        try:
-            await kube_orchestrator.delete_job(job)
-        except Exception:
-            pass
 
 
 class TestKubeOrchestrator:
@@ -412,88 +376,6 @@ class TestKubeOrchestrator:
         finally:
             await job.delete()
 
-    @pytest.fixture
-    async def ingress(self, kube_client: KubeClient) -> AsyncIterator[Ingress]:
-        ingress_name = str(uuid.uuid4())
-        ingress = await kube_client.create_ingress(ingress_name)
-        yield ingress
-        await kube_client.delete_ingress(ingress.name)
-
-    @pytest.mark.asyncio
-    async def test_ingress(self, kube_client: KubeClient, ingress: Ingress) -> None:
-        await kube_client.add_ingress_rule(ingress.name, IngressRule(host="host1"))
-        await kube_client.add_ingress_rule(ingress.name, IngressRule(host="host2"))
-        await kube_client.add_ingress_rule(ingress.name, IngressRule(host="host3"))
-        result_ingress = await kube_client.get_ingress(ingress.name)
-        assert result_ingress == Ingress(
-            name=ingress.name,
-            rules=[
-                IngressRule(host=""),
-                IngressRule(host="host1"),
-                IngressRule(host="host2"),
-                IngressRule(host="host3"),
-            ],
-        )
-
-        await kube_client.remove_ingress_rule(ingress.name, "host2")
-        result_ingress = await kube_client.get_ingress(ingress.name)
-        assert result_ingress == Ingress(
-            name=ingress.name,
-            rules=[
-                IngressRule(host=""),
-                IngressRule(host="host1"),
-                IngressRule(host="host3"),
-            ],
-        )
-
-    @pytest.mark.asyncio
-    async def test_remove_ingress_rule(
-        self, kube_client: KubeClient, ingress: Ingress
-    ) -> None:
-        with pytest.raises(StatusException, match="NotFound"):
-            await kube_client.remove_ingress_rule(ingress.name, "unknown")
-
-    @pytest.mark.asyncio
-    async def test_delete_ingress_failure(self, kube_client: KubeClient) -> None:
-        with pytest.raises(StatusException, match="NotFound"):
-            await kube_client.delete_ingress("unknown")
-
-    @pytest.mark.asyncio
-    async def test_service(self, kube_client: KubeClient) -> None:
-        service_name = f"job-{uuid.uuid4()}"
-        service = Service(name=service_name, target_port=8080)
-        try:
-            result_service = await kube_client.create_service(service)
-            assert result_service.name == service_name
-            assert result_service.target_port == 8080
-            assert result_service.port == 80
-        finally:
-            await kube_client.delete_service(service_name)
-
-    async def _wait_for_job_service(
-        self,
-        kube_ingress_ip: str,
-        host: str,
-        job_id: str,
-        interval_s: float = 0.5,
-        max_time: float = 180,
-    ) -> None:
-        url = f"http://{kube_ingress_ip}"
-        headers = {"Host": host}
-        t0 = time.monotonic()
-        async with aiohttp.ClientSession() as client:
-            while True:
-                try:
-                    async with client.get(url, headers=headers) as response:
-                        if response.status == 200:
-                            break
-                except (OSError, aiohttp.ClientError) as exc:
-                    print(exc)
-                await asyncio.sleep(max(interval_s, time.monotonic() - t0))
-                if time.monotonic() - t0 > max_time:
-                    pytest.fail(f"Failed to connect to job service {job_id}")
-                interval_s *= 1.5
-
     @pytest.mark.asyncio
     async def test_job_with_exposed_http_server_no_job_name(
         self,
@@ -501,6 +383,7 @@ class TestKubeOrchestrator:
         kube_orchestrator: KubeOrchestrator,
         kube_ingress_ip: str,
         kube_client: KubeClient,
+        wait_for_job_service: Callable[..., Awaitable[None]],
     ) -> None:
         container = Container(
             image="python",
@@ -516,7 +399,7 @@ class TestKubeOrchestrator:
 
             assert job.http_host_named is None
 
-            await self._wait_for_job_service(
+            await wait_for_job_service(
                 kube_ingress_ip, host=job.http_host, job_id=job.id
             )
 
@@ -535,6 +418,7 @@ class TestKubeOrchestrator:
         kube_orchestrator: KubeOrchestrator,
         kube_ingress_ip: str,
         kube_client: KubeClient,
+        wait_for_job_service: Callable[..., Awaitable[None]],
     ) -> None:
         container = Container(
             image="python",
@@ -545,7 +429,7 @@ class TestKubeOrchestrator:
         job = MyJob(
             orchestrator=kube_orchestrator,
             job_request=JobRequest.create(container),
-            name=f"test-job-name-{random_str()}",
+            name=f"job-{uuid.uuid4().hex[:5]}",
             owner="owner",
         )
         try:
@@ -554,9 +438,7 @@ class TestKubeOrchestrator:
 
             for host in [job.http_host, job.http_host_named]:
                 assert host
-                await self._wait_for_job_service(
-                    kube_ingress_ip, host=host, job_id=job.id
-                )
+                await wait_for_job_service(kube_ingress_ip, host=host, job_id=job.id)
                 ingress = await kube_client.get_ingress(kube_config.jobs_ingress_name)
                 assert ingress.find_rule_index_by_host(host) >= 0
 
@@ -575,6 +457,7 @@ class TestKubeOrchestrator:
         kube_orchestrator: KubeOrchestrator,
         kube_ingress_ip: str,
         kube_client: KubeClient,
+        wait_for_job_service: Callable[..., Awaitable[None]],
     ) -> None:
         container = Container(
             image="python",
@@ -590,7 +473,7 @@ class TestKubeOrchestrator:
 
             assert job.http_host_named is None
 
-            await self._wait_for_job_service(
+            await wait_for_job_service(
                 kube_ingress_ip, host=job.http_host, job_id=job.id
             )
 
@@ -609,6 +492,7 @@ class TestKubeOrchestrator:
         kube_orchestrator: KubeOrchestrator,
         kube_ingress_ip: str,
         kube_client: KubeClient,
+        wait_for_job_service: Callable[..., Awaitable[None]],
     ) -> None:
         container = Container(
             image="python",
@@ -619,7 +503,7 @@ class TestKubeOrchestrator:
         job = MyJob(
             orchestrator=kube_orchestrator,
             job_request=JobRequest.create(container),
-            name=f"test-job-name-{random_str()}",
+            name=f"job-{uuid.uuid4().hex[:5]}",
             owner="owner",
         )
         try:
@@ -627,7 +511,7 @@ class TestKubeOrchestrator:
 
             for http_host in [job.http_host, job.http_host_named]:
                 assert http_host
-                await self._wait_for_job_service(
+                await wait_for_job_service(
                     kube_ingress_ip, host=http_host, job_id=job.id
                 )
 
@@ -685,13 +569,14 @@ class TestKubeOrchestrator:
         create_client_job: Callable[[str], MyJob],
         kube_ingress_ip: str,
         delete_job_later: Callable[[Job], Awaitable[None]],
+        wait_for_job_service: Callable[..., Awaitable[None]],
     ) -> None:
         server_job = create_server_job()
         await delete_job_later(server_job)
         await server_job.start()
         server_hostname = server_job.internal_hostname
         assert server_hostname
-        await self._wait_for_job_service(
+        await wait_for_job_service(
             kube_ingress_ip, host=server_job.http_host, job_id=server_job.id
         )
 
@@ -708,13 +593,14 @@ class TestKubeOrchestrator:
         create_client_job: Callable[[str], MyJob],
         kube_ingress_ip: str,
         delete_job_later: Callable[[Job], Awaitable[None]],
+        wait_for_job_service: Callable[..., Awaitable[None]],
     ) -> None:
         server_job = create_server_job()
         await delete_job_later(server_job)
         await server_job.start()
 
         http_host = server_job.http_host
-        await self._wait_for_job_service(
+        await wait_for_job_service(
             kube_ingress_ip, host=http_host, job_id=server_job.id
         )
         client_job = create_client_job(http_host)
@@ -732,6 +618,7 @@ class TestKubeOrchestrator:
         create_client_job: Callable[[str], MyJob],
         kube_ingress_ip: str,
         delete_job_later: Callable[[Job], Awaitable[None]],
+        wait_for_job_service: Callable[..., Awaitable[None]],
     ) -> None:
         server_job_name = f"server-job-{random_str()}"
         server_job = create_server_job(job_name=server_job_name)
@@ -739,7 +626,7 @@ class TestKubeOrchestrator:
         await server_job.start()
 
         http_host = server_job.http_host
-        await self._wait_for_job_service(
+        await wait_for_job_service(
             kube_ingress_ip, host=http_host, job_id=server_job.id
         )
         client_job = create_client_job(http_host)
@@ -749,7 +636,7 @@ class TestKubeOrchestrator:
 
         assert server_job.http_host_named
         http_host = server_job.http_host_named
-        await self._wait_for_job_service(
+        await wait_for_job_service(
             kube_ingress_ip, host=http_host, job_id=server_job.id
         )
         client_job = create_client_job(http_host)
@@ -798,6 +685,26 @@ class TestKubeOrchestrator:
         await client_job.start()
         assert self.wait_for_success(client_job)
 
+
+class TestLogReader:
+    async def _consume_log_reader(
+        self, log_reader: LogReader, chunk_size: int = -1
+    ) -> bytes:
+        istream = io.BytesIO()
+        try:
+            async with log_reader:
+                while True:
+                    chunk = await log_reader.read(chunk_size)
+                    if not chunk:
+                        break
+                    assert chunk_size < 0 or len(chunk) <= chunk_size
+                    istream.write(chunk)
+        except asyncio.CancelledError:
+            pass
+        istream.flush()
+        istream.seek(0)
+        return istream.read()
+
     @pytest.mark.asyncio
     async def test_job_pod_labels_and_network_policy(
         self,
@@ -831,423 +738,6 @@ class TestKubeOrchestrator:
         assert raw_policy["spec"]["podSelector"]["matchLabels"] == {
             "platform.neuromation.io/user": job.owner
         }
-
-
-@pytest.fixture
-async def delete_pod_later(
-    kube_client: KubeClient
-) -> AsyncIterator[Callable[[PodDescriptor], Awaitable[None]]]:
-    pods = []
-
-    async def _add_pod(pod: PodDescriptor) -> None:
-        pods.append(pod)
-
-    yield _add_pod
-
-    for pod in pods:
-        try:
-            await kube_client.delete_pod(pod.name)
-        except Exception:
-            pass
-
-
-class TestKubeClient:
-    @pytest.mark.asyncio
-    async def test_wait_pod_is_running_not_found(self, kube_client: KubeClient) -> None:
-        with pytest.raises(JobNotFoundException):
-            await kube_client.wait_pod_is_running(pod_name="unknown")
-
-    @pytest.mark.asyncio
-    async def test_wait_pod_is_running_timed_out(
-        self,
-        kube_config: KubeConfig,
-        kube_client: KubeClient,
-        delete_pod_later: Callable[[PodDescriptor], Awaitable[None]],
-    ) -> None:
-        container = Container(
-            image="ubuntu",
-            command="true",
-            resources=ContainerResources(cpu=0.1, memory_mb=128),
-        )
-        job_request = JobRequest.create(container)
-        pod = PodDescriptor.from_job_request(
-            kube_config.create_storage_volume(), job_request
-        )
-        await delete_pod_later(pod)
-        await kube_client.create_pod(pod)
-        with pytest.raises(asyncio.TimeoutError):
-            await kube_client.wait_pod_is_running(pod_name=pod.name, timeout_s=0.1)
-
-    @pytest.mark.asyncio
-    async def test_wait_pod_is_running(
-        self,
-        kube_config: KubeConfig,
-        kube_client: KubeClient,
-        delete_pod_later: Callable[[PodDescriptor], Awaitable[None]],
-    ) -> None:
-        container = Container(
-            image="ubuntu",
-            command="true",
-            resources=ContainerResources(cpu=0.1, memory_mb=128),
-        )
-        job_request = JobRequest.create(container)
-        pod = PodDescriptor.from_job_request(
-            kube_config.create_storage_volume(), job_request
-        )
-        await delete_pod_later(pod)
-        await kube_client.create_pod(pod)
-        await kube_client.wait_pod_is_running(pod_name=pod.name, timeout_s=60.0)
-        pod_status = await kube_client.get_pod_status(pod.name)
-        assert pod_status.phase in ("Running", "Succeeded")
-
-    @pytest.mark.asyncio
-    async def test_create_log_stream_not_found(self, kube_client: KubeClient) -> None:
-        with pytest.raises(KubeClientException):
-            async with kube_client.create_pod_container_logs_stream(
-                pod_name="unknown", container_name="unknown"
-            ):
-                pass
-
-    @pytest.mark.asyncio
-    async def test_create_log_stream_creating(
-        self,
-        kube_config: KubeConfig,
-        kube_client: KubeClient,
-        delete_pod_later: Callable[[PodDescriptor], Awaitable[None]],
-    ) -> None:
-        container = Container(
-            image="ubuntu",
-            command="true",
-            resources=ContainerResources(cpu=0.1, memory_mb=128),
-        )
-        job_request = JobRequest.create(container)
-        pod = PodDescriptor.from_job_request(
-            kube_config.create_storage_volume(), job_request
-        )
-        await delete_pod_later(pod)
-        await kube_client.create_pod(pod)
-
-        async with timeout(5.0):
-            while True:
-                try:
-                    stream_cm = kube_client.create_pod_container_logs_stream(
-                        pod_name=pod.name, container_name=pod.name
-                    )
-                    with pytest.raises(KubeClientException, match="ContainerCreating"):
-                        async with stream_cm:
-                            pass
-                    break
-                except AssertionError as exc:
-                    if "Pattern" not in str(exc):
-                        raise
-                await asyncio.sleep(0.1)
-
-    @pytest.mark.asyncio
-    async def test_create_log_stream(
-        self,
-        kube_config: KubeConfig,
-        kube_client: KubeClient,
-        delete_pod_later: Callable[[PodDescriptor], Awaitable[None]],
-    ) -> None:
-        container = Container(
-            image="ubuntu",
-            command="true",
-            resources=ContainerResources(cpu=0.1, memory_mb=128),
-        )
-        job_request = JobRequest.create(container)
-        pod = PodDescriptor.from_job_request(
-            kube_config.create_storage_volume(), job_request
-        )
-        await delete_pod_later(pod)
-        await kube_client.create_pod(pod)
-        await kube_client.wait_pod_is_running(pod_name=pod.name, timeout_s=60.0)
-        stream_cm = kube_client.create_pod_container_logs_stream(
-            pod_name=pod.name, container_name=pod.name
-        )
-        async with stream_cm as stream:
-            payload = await stream.read()
-            assert payload == b""
-
-    @pytest.mark.asyncio
-    async def test_create_docker_secret_non_existent_namespace(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        name = str(uuid.uuid4())
-        docker_secret = DockerRegistrySecret(
-            name=name,
-            namespace=name,
-            username="testuser",
-            password="testpassword",
-            email="testuser@example.com",
-            registry_server="registry.example.com",
-        )
-
-        with pytest.raises(StatusException, match="NotFound"):
-            await kube_client.create_docker_secret(docker_secret)
-
-    @pytest.mark.asyncio
-    async def test_create_docker_secret_already_exists(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        name = str(uuid.uuid4())
-        docker_secret = DockerRegistrySecret(
-            name=name,
-            namespace=kube_config.namespace,
-            username="testuser",
-            password="testpassword",
-            email="testuser@example.com",
-            registry_server="registry.example.com",
-        )
-
-        try:
-            await kube_client.create_docker_secret(docker_secret)
-
-            with pytest.raises(StatusException, match="AlreadyExists"):
-                await kube_client.create_docker_secret(docker_secret)
-        finally:
-            await kube_client.delete_secret(name, kube_config.namespace)
-
-    @pytest.mark.asyncio
-    async def test_update_docker_secret_already_exists(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        name = str(uuid.uuid4())
-        docker_secret = DockerRegistrySecret(
-            name=name,
-            namespace=kube_config.namespace,
-            username="testuser",
-            password="testpassword",
-            email="testuser@example.com",
-            registry_server="registry.example.com",
-        )
-
-        try:
-            await kube_client.create_docker_secret(docker_secret)
-            await kube_client.update_docker_secret(docker_secret)
-        finally:
-            await kube_client.delete_secret(name, kube_config.namespace)
-
-    @pytest.mark.asyncio
-    async def test_update_docker_secret_non_existent(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        name = str(uuid.uuid4())
-        docker_secret = DockerRegistrySecret(
-            name=name,
-            namespace=kube_config.namespace,
-            username="testuser",
-            password="testpassword",
-            email="testuser@example.com",
-            registry_server="registry.example.com",
-        )
-
-        with pytest.raises(StatusException, match="NotFound"):
-            await kube_client.update_docker_secret(docker_secret)
-
-    @pytest.mark.asyncio
-    async def test_update_docker_secret_create_non_existent(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        name = str(uuid.uuid4())
-        docker_secret = DockerRegistrySecret(
-            name=name,
-            namespace=kube_config.namespace,
-            username="testuser",
-            password="testpassword",
-            email="testuser@example.com",
-            registry_server="registry.example.com",
-        )
-
-        await kube_client.update_docker_secret(docker_secret, create_non_existent=True)
-        await kube_client.update_docker_secret(docker_secret)
-
-    @pytest.fixture
-    async def delete_network_policy_later(
-        self, kube_client: KubeClient
-    ) -> AsyncIterator[Callable[[str], Awaitable[None]]]:
-        names = []
-
-        async def _add_name(name: str) -> None:
-            names.append(name)
-
-        yield _add_name
-
-        for name in names:
-            try:
-                await kube_client.delete_network_policy(name)
-            except Exception:
-                pass
-
-    @pytest.mark.asyncio
-    async def test_create_default_network_policy(
-        self,
-        kube_config: KubeConfig,
-        kube_client: KubeClient,
-        delete_network_policy_later: Callable[[str], Awaitable[None]],
-    ) -> None:
-        name = str(uuid.uuid4())
-        await delete_network_policy_later(name)
-        payload = await kube_client.create_default_network_policy(
-            name, {"testlabel": name}, namespace_name=kube_config.namespace
-        )
-        assert payload["metadata"]["name"] == name
-
-    @pytest.mark.asyncio
-    async def test_create_default_network_policy_twice(
-        self,
-        kube_config: KubeConfig,
-        kube_client: KubeClient,
-        delete_network_policy_later: Callable[[str], Awaitable[None]],
-    ) -> None:
-        name = str(uuid.uuid4())
-        await delete_network_policy_later(name)
-        payload = await kube_client.create_default_network_policy(
-            name, {"testlabel": name}, namespace_name=kube_config.namespace
-        )
-        assert payload["metadata"]["name"] == name
-        with pytest.raises(AlreadyExistsException):
-            await kube_client.create_default_network_policy(
-                name, {"testlabel": name}, namespace_name=kube_config.namespace
-            )
-
-    @pytest.mark.asyncio
-    async def test_get_network_policy_not_found(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        name = str(uuid.uuid4())
-        with pytest.raises(StatusException, match="NotFound"):
-            await kube_client.get_network_policy(name)
-
-    @pytest.mark.asyncio
-    async def test_delete_network_policy_not_found(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        name = str(uuid.uuid4())
-        with pytest.raises(StatusException, match="NotFound"):
-            await kube_client.delete_network_policy(name)
-
-    @pytest.mark.asyncio
-    async def test_get_pod_events(
-        self,
-        kube_config: KubeConfig,
-        kube_client: MyKubeClient,
-        delete_pod_later: Callable[[PodDescriptor], Awaitable[None]],
-    ) -> None:
-        container = Container(
-            image="ubuntu",
-            command="true",
-            resources=ContainerResources(cpu=0.1, memory_mb=128),
-        )
-        job_request = JobRequest.create(container)
-        pod = PodDescriptor.from_job_request(
-            kube_config.create_storage_volume(), job_request
-        )
-        await delete_pod_later(pod)
-        await kube_client.create_pod(pod)
-        await kube_client.wait_pod_is_terminated(pod.name)
-
-        events = await kube_client.get_pod_events(pod.name, kube_config.namespace)
-
-        assert events
-        for event in events:
-            involved_object = event.involved_object
-            assert involved_object["kind"] == "Pod"
-            assert involved_object["namespace"] == kube_config.namespace
-            assert involved_object["name"] == pod.name
-
-    @pytest.mark.asyncio
-    async def test_get_pod_events_empty(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        pod_name = str(uuid.uuid4())
-        events = await kube_client.get_pod_events(pod_name, kube_config.namespace)
-
-        assert not events
-
-    @pytest.mark.asyncio
-    async def test_get_pod_container_stats(
-        self,
-        kube_config: KubeConfig,
-        kube_client: KubeClient,
-        delete_pod_later: Callable[[PodDescriptor], Awaitable[None]],
-    ) -> None:
-        command = 'bash -c "for i in {1..5}; do echo $i; sleep 1; done"'
-        container = Container(
-            image="ubuntu",
-            command=command,
-            resources=ContainerResources(cpu=0.1, memory_mb=128),
-        )
-        job_request = JobRequest.create(container)
-        pod = PodDescriptor.from_job_request(
-            kube_config.create_storage_volume(), job_request
-        )
-        await delete_pod_later(pod)
-        await kube_client.create_pod(pod)
-        await kube_client.wait_pod_is_running(pod_name=pod.name, timeout_s=60.0)
-
-        pod_metrics = []
-        while True:
-            stats = await kube_client.get_pod_container_stats(pod.name, pod.name)
-            if stats:
-                pod_metrics.append(stats)
-            else:
-                break
-            await asyncio.sleep(1)
-
-        assert pod_metrics
-        assert pod_metrics[0] == PodContainerStats(cpu=mock.ANY, memory=mock.ANY)
-        assert pod_metrics[0].cpu >= 0.0
-        assert pod_metrics[0].memory > 0.0
-
-    @pytest.mark.asyncio
-    async def test_get_pod_container_stats_no_pod(
-        self, kube_config: KubeConfig, kube_client: KubeClient
-    ) -> None:
-        pod_name = str(uuid.uuid4())
-        with pytest.raises(JobNotFoundException):
-            await kube_client.get_pod_container_stats(pod_name, pod_name)
-
-    @pytest.mark.asyncio
-    async def test_get_pod_container_stats_not_scheduled_yet(
-        self,
-        kube_config: KubeConfig,
-        kube_client: KubeClient,
-        delete_pod_later: Callable[[PodDescriptor], Awaitable[None]],
-    ) -> None:
-        container = Container(
-            image="ubuntu",
-            command="true",
-            resources=ContainerResources(cpu=100, memory_mb=128),
-        )
-        job_request = JobRequest.create(container)
-        pod = PodDescriptor.from_job_request(
-            kube_config.create_storage_volume(), job_request
-        )
-        await delete_pod_later(pod)
-        await kube_client.create_pod(pod)
-
-        stats = await kube_client.get_pod_container_stats(pod.name, pod.name)
-        assert stats is None
-
-
-class TestLogReader:
-    async def _consume_log_reader(
-        self, log_reader: LogReader, chunk_size: int = -1
-    ) -> bytes:
-        istream = io.BytesIO()
-        try:
-            async with log_reader:
-                while True:
-                    chunk = await log_reader.read(chunk_size)
-                    if not chunk:
-                        break
-                    assert chunk_size < 0 or len(chunk) <= chunk_size
-                    istream.write(chunk)
-        except asyncio.CancelledError:
-            pass
-        istream.flush()
-        istream.seek(0)
-        return istream.read()
 
     @pytest.mark.asyncio
     async def test_read_instantly_succeeded(
